+++ comm/model/playerinfo.go
@@ -7,7 +7,7 @@ package model
 type PlayerInfo struct {
 	GameId     string `json:"game_id"`
 	PlayerId   string `json:"player_id"`
-	IsVip      int64  `json:"is_vip"`
+	IsVip      int32  `json:"is_vip"`
 	CsId       string `json:"cs_id"`
 	ConnectTs  int64  `json:"connect_ts"`
 	LastChatTs int64  `json:"last_chat_ts"`
